+++ src/main/java/lotto/util/Validator.java
@@ -1,4 +1,6 @@
-package lotto;
+package lotto.util;
+
+import lotto.Env;
 
 import java.util.LinkedList;
 import java.util.List;
